requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-import-export==3.0.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.6
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schedule==1.1.0
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
